refactor(gui): modularize ResultView into dedicated package
Extract result drawing logic (amplicon drawing, primer drawing,
and diagram panel) from ResultView into separate components in a
new result/ package.
