+++ src/amplifyp/gui/views/result/primer_drawing.py
@@ -1,0 +1,367 @@
+# Copyright (C) 2026 Fufu Fang
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+"""Primer drawing class and context card helpers for the ResultView."""
+
+from collections.abc import Callable
+from typing import Any
+
+import flet as ft
+import flet.canvas as cv
+
+from amplifyp.dna import DNA, DNADirection, DNAType
+from amplifyp.gui.settings import GUIColors
+
+
+class DrawnPrimer:
+    """Class representing a primer drawn on the PCR result diagram canvas."""
+
+    def __init__(
+        self,
+        name: str,
+        index: int,
+        conf: Any,
+        var: Any,
+        S: float,
+        target_length: int,
+        t_width: float,
+        h_margin: float,
+        v_target: float,
+        settings: Any,
+        on_click: Callable[[], None],
+    ) -> None:
+        """Initialize the DrawnPrimer."""
+        self.name = name
+        self.index = index
+        self.conf = conf
+        self.var = var
+        self.S = S
+        self.target_length = target_length
+        self.t_width = t_width
+        self.h_margin = h_margin
+        self.v_target = v_target
+        self.settings = settings
+        self.on_click = on_click
+
+        # Calculate coordinates
+        self.x_pos = (
+            h_margin + (index / target_length * t_width)
+            if target_length
+            else h_margin
+        )
+        self.direction = var.direction
+
+    def draw(self, canvas: cv.Canvas, stack: ft.Stack) -> None:
+        """Draw the primer indicator elements onto the canvas and stack."""
+        if self.direction == DNADirection.FWD:
+            # Draw FWD primers (blue, float above baseline, pointing down)
+            # Connector line from baseline to floating tip
+            canvas.shapes.append(
+                cv.Path(
+                    [
+                        cv.Path.MoveTo(self.x_pos, self.v_target),
+                        cv.Path.LineTo(self.x_pos, self.v_target - 25),
+                    ],
+                    paint=ft.Paint(
+                        color=GUIColors.FWD_PRIMER,
+                        style=ft.PaintingStyle.STROKE,
+                        stroke_width=1.0,
+                    ),
+                )
+            )
+            # Down-pointing triangle of size S:
+            # Tip: (self.x_pos, self.v_target - 25)
+            # Top-Left: (self.x_pos - S/2, self.v_target - 25 - S)
+            # Top-Right: (self.x_pos + S/2, self.v_target - 25 - S)
+            canvas.shapes.append(
+                cv.Path(
+                    [
+                        cv.Path.MoveTo(self.x_pos, self.v_target - 25),
+                        cv.Path.LineTo(
+                            self.x_pos - self.S / 2, self.v_target - 25 - self.S
+                        ),
+                        cv.Path.LineTo(
+                            self.x_pos + self.S / 2, self.v_target - 25 - self.S
+                        ),
+                        cv.Path.Close(),
+                    ],
+                    paint=ft.Paint(
+                        color=GUIColors.FWD_PRIMER,
+                        style=ft.PaintingStyle.FILL,
+                    ),
+                )
+            )
+            # Add rotated label in the stack
+            stack.controls.append(
+                ft.Text(
+                    self.name,
+                    color=GUIColors.FWD_PRIMER,
+                    size=self.settings.get("font_size_map_primer", 13),
+                    weight=ft.FontWeight.BOLD,
+                    left=self.x_pos - 15,
+                    top=self.v_target - 25 - self.S - 38,
+                    rotate=ft.Rotate(-1.5708),
+                )
+            )
+            # Click overlay
+            stack.controls.append(
+                ft.GestureDetector(
+                    mouse_cursor=ft.MouseCursor.CLICK,
+                    on_tap=lambda _: self.on_click(),
+                    content=ft.Container(
+                        bgcolor=GUIColors.TRANSPARENT,
+                        width=20,
+                        height=25 + self.S,
+                    ),
+                    left=self.x_pos - 10,
+                    top=self.v_target - 25 - self.S,
+                )
+            )
+        else:
+            # Draw REV primers (red, float below baseline, pointing up)
+            # Connector line from baseline to floating tip
+            canvas.shapes.append(
+                cv.Path(
+                    [
+                        cv.Path.MoveTo(self.x_pos, self.v_target),
+                        cv.Path.LineTo(self.x_pos, self.v_target + 25),
+                    ],
+                    paint=ft.Paint(
+                        color=GUIColors.REV_PRIMER,
+                        style=ft.PaintingStyle.STROKE,
+                        stroke_width=1.0,
+                    ),
+                )
+            )
+            # Up-pointing triangle of size S:
+            # Tip: (self.x_pos, self.v_target + 25)
+            # Bottom-Left: (self.x_pos - S/2, self.v_target + 25 + S)
+            # Bottom-Right: (self.x_pos + S/2, self.v_target + 25 + S)
+            canvas.shapes.append(
+                cv.Path(
+                    [
+                        cv.Path.MoveTo(self.x_pos, self.v_target + 25),
+                        cv.Path.LineTo(
+                            self.x_pos - self.S / 2, self.v_target + 25 + self.S
+                        ),
+                        cv.Path.LineTo(
+                            self.x_pos + self.S / 2, self.v_target + 25 + self.S
+                        ),
+                        cv.Path.Close(),
+                    ],
+                    paint=ft.Paint(
+                        color=GUIColors.REV_PRIMER,
+                        style=ft.PaintingStyle.FILL,
+                    ),
+                )
+            )
+            # Add rotated label in the stack
+            stack.controls.append(
+                ft.Text(
+                    self.name,
+                    color=GUIColors.REV_LABEL,
+                    size=self.settings.get("font_size_map_primer", 13),
+                    weight=ft.FontWeight.BOLD,
+                    left=self.x_pos - 15,
+                    top=self.v_target + 25 + self.S + 10,
+                    rotate=ft.Rotate(-1.5708),
+                )
+            )
+            # Click overlay
+            stack.controls.append(
+                ft.GestureDetector(
+                    mouse_cursor=ft.MouseCursor.CLICK,
+                    on_tap=lambda _: self.on_click(),
+                    content=ft.Container(
+                        bgcolor=GUIColors.TRANSPARENT,
+                        width=20,
+                        height=25 + self.S,
+                    ),
+                    left=self.x_pos - 10,
+                    top=self.v_target,
+                )
+            )
+
+
+def get_template_substring(template: DNA, start: int, length: int) -> str:
+    """Get substring of template DNA, supporting circular wrapping."""
+    N_len = len(template)
+    if template.type == DNAType.CIRCULAR:
+        return "".join(
+            template.seq[i % N_len] for i in range(start, start + length)
+        )
+    else:
+        return "".join(
+            template.seq[i] if 0 <= i < N_len else "-"
+            for i in range(start, start + length)
+        )
+
+
+def format_context_lines(
+    primer_name: str,
+    padded_idx: int,
+    conf: Any,
+    origin: Any,
+    L: int,
+    N: int,
+    direction: DNADirection,
+) -> tuple[str, str, str]:
+    """Format context alignment lines for binding site context map."""
+    if direction == DNADirection.FWD:
+        start_genomic = (padded_idx - L) % N
+        primer_display_seq = conf.primer.seq
+        primer_label = f"{primer_name} (Forward)"
+        primer_ends = ("5'", "3'")
+        strength_display = origin.binding_strength_str[::-1]
+    else:
+        start_genomic = padded_idx % N
+        primer_display_seq = conf.primer.seq[::-1]
+        primer_label = f"{primer_name} (Reverse)"
+        primer_ends = ("3'", "5'")
+        strength_display = origin.binding_strength_str
+
+    end_genomic = start_genomic + L
+    start_num_str = str(((start_genomic) % N) + 1)
+    end_num_str = str(((end_genomic - 1) % N) + 1)
+
+    # Construct primer line:
+    primer_line = (
+        f"{primer_label:<29}"
+        f"{primer_ends[0]}-{primer_display_seq}-{primer_ends[1]}"
+    )
+
+    # Construct strength line:
+    bonds_line = f"{' ' * 12}{' ' * 20}{strength_display}"
+
+    # Construct arrows:
+    arrow_line = f"{' ' * 12}{' ' * 20}↓{' ' * (L - 2)}↓"
+
+    # Construct numbers:
+    num_line_chars = [" "] * (12 + 20 + L + 20)
+    col1 = 12 + 20 - len(start_num_str) // 2
+    for idx, char in enumerate(start_num_str):
+        num_line_chars[col1 + idx] = char
+    col2 = (12 + 20 + L - 1) - len(end_num_str) // 2
+    for idx, char in enumerate(end_num_str):
+        num_line_chars[col2 + idx] = char
+    top_line = "".join(num_line_chars).rstrip()
+
+    # Combined top line:
+    top_line = f"{top_line}\n{arrow_line}"
+
+    upstream_seq = get_template_substring(conf.template, start_genomic - 20, 20)
+    binding_seq = get_template_substring(conf.template, start_genomic, L)
+    downstream_seq = get_template_substring(
+        conf.template, start_genomic + L, 20
+    )
+
+    context_line_prefix = "Context  "
+    bottom_line = (
+        f"{bonds_line}\n"
+        f"{context_line_prefix}5'-{upstream_seq}"
+        f"{binding_seq}{downstream_seq}-3'"
+    )
+
+    return top_line, primer_line, bottom_line
+
+
+class ReplicationContextCard(ft.Card):  # type: ignore[misc]
+    """Card showing visually aligned replication context map."""
+
+    def __init__(
+        self,
+        primer_name: str,
+        padded_idx: int,
+        conf: Any,
+        var: Any,
+        settings: Any,
+        dismiss_callback: Callable[[ft.Card], None],
+    ) -> None:
+        """Initialize the ReplicationContextCard."""
+        super().__init__()
+        self._card_id = f"context_{primer_name}_{padded_idx}"
+
+        def remove_card(e: Any) -> None:
+            dismiss_callback(self)
+
+        origin = conf.origin(var)
+        if origin is None:
+            self.content = ft.Text("Error: Replication origin not found")
+            return
+
+        primer_type = (
+            "Forward Primer"
+            if var.direction == DNADirection.FWD
+            else "Reverse Primer"
+        )
+        L = len(conf.primer)
+        N = len(conf.template)
+
+        from amplifyp.gui.util import create_overlapped_sequence_view
+
+        top_line, primer_line, bottom_line = format_context_lines(
+            primer_name=primer_name,
+            padded_idx=padded_idx,
+            conf=conf,
+            origin=origin,
+            L=L,
+            N=N,
+            direction=var.direction,
+        )
+
+        font_family = settings.get("font_family", "Roboto Mono")
+        font_size = settings.get("font_size_default", 14)
+        diagram_text = create_overlapped_sequence_view(
+            top_line=top_line,
+            mid_line=primer_line,
+            bottom_line=bottom_line,
+            font_family=font_family,
+            font_size=font_size,
+        )
+
+        self.content = ft.Container(
+            padding=10,
+            content=ft.Column(
+                [
+                    ft.Row(
+                        [
+                            ft.Text(
+                                f"Context Map - {primer_name} "
+                                f"({primer_type}) Binding Site",
+                                weight=ft.FontWeight.BOLD,
+                                size=settings.get("font_size_subheader", 16),
+                            ),
+                            ft.IconButton(
+                                icon=ft.Icons.CLOSE,
+                                icon_size=18,
+                                tooltip="Dismiss",
+                                on_click=remove_card,
+                            ),
+                        ],
+                        alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
+                    ),
+                    ft.Container(
+                        content=ft.Row(
+                            [diagram_text],
+                            scroll=ft.ScrollMode.ALWAYS,
+                        ),
+                        padding=12,
+                        border_radius=6,
+                        border=ft.Border.all(1, GUIColors.OUTLINE_VARIANT),
+                    ),
+                ],
+                tight=True,
+            ),
+        )
